requirement.txt
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2023.3.post1
ruamel.yaml.clib==0.2.8
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
